finished

+++ 16_subseq/maxSeq.c
@@ -4,10 +4,10 @@
 size_t maxSeq(int * array, size_t n){
   if (n==0) return 0;
 
-  int count=1;
-  int pcount=1;
-  int current;
-  int perivious=array[0];
+  size_t count=1;
+  size_t pcount=1;
+  size_t current;
+  size_t perivious=array[0];
   for (size_t i=0;i<n;i++){
     current=array[i];
     if (current>perivious){
